Decode qr code string

+++ Meshtastic/MeshtasticApp.swift
@@ -42,8 +42,7 @@ struct MeshtasticAppleApp: App {
 				}
 			}
 			.sheet(isPresented: $saveChannels) {
-				
-				let channelSettingsString = incomingUrl?.absoluteString
+								
 				SaveChannelQRCode(channelHash: channelSettings ?? "Empty Channel URL")
 					.presentationDetents([.medium, .large])
 					.presentationDragIndicator(.visible)
@@ -60,6 +59,7 @@ struct MeshtasticAppleApp: App {
 					saveChannels = true
 					print("User wants to open a Channel Settings URL: \(incomingUrl?.absoluteString ?? "No QR Code Link")")
 				} else {
+					saveChannels = false
 					print("User wants to import a MBTILES offline map file: \(incomingUrl?.absoluteString ?? "No Tiles link")")
 				}
 
@@ -69,26 +69,28 @@ struct MeshtasticAppleApp: App {
 				let documentsDirectory = fileManager.urls(for: .documentDirectory, in: .userDomainMask).first!
 				let destination = documentsDirectory.appendingPathComponent("offline_map.mbtiles", isDirectory: false)
 				
-				//do we need to delete an old one?
-				if (fileManager.fileExists(atPath: destination.path)) {
-					print("ℹ️ Found an old map file.  Deleting it")
-					try? fileManager.removeItem(atPath: destination.path)
-				}
-				
-				do {
-					try fileManager.copyItem(at: url, to: destination)
-				} catch {
-					print("Copy MB Tile file failed. Error: \(error)")
-				}
-				
-				if (fileManager.fileExists(atPath: destination.path)) {
-					print("ℹ️ Saved the map file")
+				if !saveChannels {
+					//do we need to delete an old one?
+					if (fileManager.fileExists(atPath: destination.path)) {
+						print("ℹ️ Found an old map file.  Deleting it")
+						try? fileManager.removeItem(atPath: destination.path)
+					}
 					
-					//need to tell the map view that it needs to update and try loading the new overlay
-					UserDefaults.standard.set(Date().timeIntervalSince1970, forKey: "lastUpdatedLocalMapFile")
+					do {
+						try fileManager.copyItem(at: url, to: destination)
+					} catch {
+						print("Copy MB Tile file failed. Error: \(error)")
+					}
 					
-				} else {
-					print("💥 Didn't save the map file")
+					if (fileManager.fileExists(atPath: destination.path)) {
+						print("ℹ️ Saved the map file")
+						
+						//need to tell the map view that it needs to update and try loading the new overlay
+						UserDefaults.standard.set(Date().timeIntervalSince1970, forKey: "lastUpdatedLocalMapFile")
+						
+					} else {
+						print("💥 Didn't save the map file")
+					}
 				}
 			})
 		}
